Attendence Added

+++ Tactsoft.Core/ViewModel/AttendenceModel.cs
@@ -11,6 +11,7 @@ namespace Tactsoft.Core.ViewModel
             AttendenceList = new Collection<AttendenceList>();
             AttendenceDate = DateTime.Now;
         }
+
         [Required]
         [DisplayName("Department")]
         public int DepartmentId { get; set; }
@@ -24,8 +25,8 @@ namespace Tactsoft.Core.ViewModel
 
     public class AttendenceList
     {
-        public int AttendenceId { get; set; }
-        public int EmployeeId { get; set; }
+        public long AttendenceId { get; set; }
+        public long EmployeeId { get; set; }
         public string Name { get; set; }
         public Boolean IsPresent { get; set; }
     }
